refactor: split client key handlers
